Move the media tests to the code they pin
They sat in the room loop's test module while the functions they check,
the PCM framing, the playout queue and the JPEG encoder, live one file
down, so a reader of that file saw untested code that was covered next
door. Their fixture goes with them, at module scope rather than inside
the test module, because the loop's own tests build one too. Four tests
that had been sitting outside any test module, running as livekit
rather than livekit::tests, move inside it.

The agent-state check in the browser suite now reads the whole runtime
rather than the root file. It counts publish sites, and a state declared
in a sibling module would have gone uncounted, which is the direction a
pin must never fail.
